# backend/src/routes/v1/swagger_api_v1.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Anonymous Voting App - Backend API",
        "description": "This is a Swagger documentation for Anonymous Voting App",
        "license": {
            "name": "MIT License",
            "url": "https://github.com/Anonymous-Voting-App/Anonymous-Voting-App/blob/dev/LICENSE"
        },
        "version": "1.0.0"
    },
    "externalDocs": {
        "description": "Repository",
        "url": "https://github.com/Anonymous-Voting-App/Anonymous-Voting-App"
    },
    "servers": [
        {
            "url": "https://d2l9c7s1roosdj.cloudfront.net/"
        }
    ],
    "tags": [
        {
            "name": "Poll admin",
            "description": "Adminstration endpoints for polls"
        },
        {
            "name": "Poll public",
            "description": "Public endpoints for polls"
        },
        {
            "name": "Maintenance",
            "description": "Health check"
        }
    ],
    "paths": {
        "/api/poll": {
            "post": {
                "tags": ["Poll public"],
                "description": "Creates a new poll from given information for specified user. A poll consists of questions and each question has sub-questions. The sub-questions correspond to the different choices a question has. For now, the sub-question titles have to be unique in the poll due to how the database schema is currently set up.\n\nCurrently a test user \"1eb1cfae-09e7-4456-85cd-e2edfff80544\" can be set as the owner of new polls. Descriptions are also not saved to the database currently due to the database schema. The type properties for the poll, questions and sub-questions need to be given. The type property for the poll currently does nothing so it can be set to an empty string.\n",
                "responses": {
                    "201": {
                        "description": "Poll created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PollPrivateInfo"
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "description": "Create a new poll",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddPoll"
                            }
                        }
                    }
                }
            }
        },
        "/api/poll/{publicId}": {
            "get": {
                "tags": ["Poll public"],
                "description": "Fetches a poll's public information using a publicId of the poll.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "publicId",
                        "description": "publicId of the poll to get.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Public information about poll.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PollPublicInfo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Poll not found"
                    }
                }
            }
        },
        "/api/poll/{publicId}/answers": {
            "get": {
                "tags": ["Poll public"],
                "description": "Get public answer data of the poll.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "publicId",
                        "description": "publicId of the poll to get.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Information about the answers to the poll.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "answers": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PrivateAnswer"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Poll not found"
                    }
                }
            },
            "post": {
                "tags": ["Poll public"],
                "description": "Answers all the poll's questions at once. All questions directly under the poll are mandatory and need to be answered. On the other hand, a sub-question of a multi-question follows the restrictions of the multi-question and thus may contain optional questions.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "publicId",
                        "description": "publicId of the poll to answer.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Whether answer was successfully given.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessMessage"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Poll not found"
                    }
                },
                "requestBody": {
                    "description": "Post an answer to a poll",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PollAnswerRequest"
                            }
                        }
                    }
                }
            }
        },
        "/api/poll/{publicId}/results": {
            "get": {
                "tags": ["Poll public"],
                "description": "Returns answer result statistics about the poll. For each question result, .answerPercentage tells the ratio of the question's .answerCount to its parent's .answerCount. The parent of a sub-question is the multi-question containing it. On the other hand, the parent of a non-sub-question is the poll itself. .answerValueStatistics contains information about the answer counts for different answer values under a question. The .percentage of these statistic objects is the ratio of the statistic object's .count to the question's .answerCount.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "publicId",
                        "description": "publicId of the poll to get results of.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Information about the poll's answer results.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PollResults"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Poll not found"
                    }
                }
            }
        },
        "/api/poll/admin/{privateId}": {
            "get": {
                "tags": ["Poll admin"],
                "description": "Fetches a poll's private information using a privateId of the poll for the admin view.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "privateId",
                        "description": "privateId of the poll to get.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Private information about poll.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PollPrivateInfo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Poll not found"
                    }
                }
            }
        },
        "/api/health": {
            "get": {
                "tags": ["Maintenance"],
                "description": "Health check of API",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "server": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/health/db": {
            "get": {
                "tags": ["Maintenance"],
                "description": "Health check of API and database connection",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "server": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "database": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "QuestionType": {
                "type": "string",
                "description": "One of: \"free\" (free-form text question), \"multi\" (question containing sub-questions), \"number\" (number question), \"scale\" (number question with a specified range of allowed values).",
                "example": "free"
            },
            "AddPoll": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Name of the poll"
                    },
                    "type": {
                        "type": "string"
                    },
                    "owner": {
                        "type": "object",
                        "properties": {
                            "accountId": {
                                "type": "string",
                                "format": "uuid"
                            }
                        }
                    },
                    "questions": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "title": {
                                    "type": "string",
                                    "example": "Title of question"
                                },
                                "description": {
                                    "type": "string",
                                    "example": "Description of question"
                                },
                                "type": {
                                    "type": "string"
                                },
                                "subQuestions": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "Title of sub-question"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Description of sub-question"
                                            },
                                            "type": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "PollPrivateInfo": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "example": "Name of the poll"
                    },
                    "publicId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "privateId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "type": {
                        "type": "string"
                    },
                    "owner": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "string",
                                "format": "uuid"
                            }
                        }
                    },
                    "questions": {
                        "type": "array",
                        "items": {
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/Question"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "subQuestions": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Question"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "answers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PublicAnswer"
                        }
                    }
                }
            },
            "PollPublicInfo": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "example": "Name of the poll"
                    },
                    "publicId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "questions": {
                        "type": "array",
                        "items": {
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/Question"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "subQuestions": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Question"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "PollResults": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "example": "Name of the poll"
                    },
                    "publicId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "questions": {
                        "type": "array",
                        "items": {
                            "oneOf": [
                                {
                                    "$ref": "#/components/schemas/MultiQuestionResult"
                                },
                                {
                                    "$ref": "#/components/schemas/QuestionResult"
                                }
                            ]
                        }
                    }
                }
            },
            "AddAnswer": {
                "properties": {
                    "questionId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "type": {
                        "type": "string",
                        "example": "multi"
                    },
                    "answer": {
                        "oneOf": [
                            {
                                "type": "object",
                                "properties": {
                                    "subQuestionIds": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "uuid"
                                        }
                                    },
                                    "answer": {
                                        "type": "array",
                                        "items": {
                                            "type": "object",
                                            "properties": {
                                                "answer": {
                                                    "type": "string",
                                                    "example": "test-text-value"
                                                }
                                            }
                                        }
                                    }
                                }
                            },
                            {
                                "type": "object",
                                "properties": {
                                    "answer": {
                                        "type": "string",
                                        "example": "test-text-value"
                                    }
                                }
                            }
                        ]
                    }
                }
            },
            "PollAnswerRequest": {
                "properties": {
                    "answers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AddAnswer"
                        }
                    }
                }
            },
            "PrivateAnswer": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/PublicAnswer"
                    },
                    {
                        "type": "object",
                        "properties": {
                            "answerer": {
                                "properties": {
                                    "id": {
                                        "type": "string",
                                        "format": "uuid"
                                    }
                                }
                            }
                        }
                    }
                ]
            },
            "PublicAnswer": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "questionId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "value": {
                        "type": "string",
                        "example": true
                    },
                    "subAnswers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PublicAnswer"
                        }
                    }
                }
            },
            "Question": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string",
                        "example": "Title of the sub-question"
                    },
                    "description": {
                        "type": "string",
                        "example": "Description of the sub-question"
                    },
                    "pollId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "type": {
                        "$ref": "#/components/schemas/QuestionType"
                    },
                    "minAnswers": {
                        "type": "number",
                        "example": 1
                    },
                    "maxAnswers": {
                        "type": "number",
                        "example": 1
                    },
                    "minValue": {
                        "type": "number",
                        "example": 1
                    },
                    "maxValue": {
                        "type": "number",
                        "example": 5
                    },
                    "step": {
                        "type": "number",
                        "example": 1
                    }
                }
            },
            "QuestionResult": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string",
                        "example": "Title of the sub-question"
                    },
                    "description": {
                        "type": "string",
                        "example": "Description of the sub-question"
                    },
                    "pollId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "type": {
                        "$ref": "#/components/schemas/QuestionType"
                    },
                    "answerCount": {
                        "type": "number",
                        "example": 1
                    },
                    "answerPercentage": {
                        "type": "number",
                        "example": 1
                    },
                    "answerValueStatistics": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AnswerValueStatistic"
                        }
                    }
                }
            },
            "MultiQuestionResult": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string",
                        "example": "Title of the sub-question"
                    },
                    "description": {
                        "type": "string",
                        "example": "Description of the sub-question"
                    },
                    "pollId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "type": {
                        "type": "string",
                        "example": "multi"
                    },
                    "answerCount": {
                        "type": "number",
                        "example": 1
                    },
                    "answerPercentage": {
                        "type": "number",
                        "example": 1
                    },
                    "subQuestions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/QuestionResult"
                        }
                    }
                }
            },
            "AnswerValueStatistic": {
                "properties": {
                    "value": {
                        "type": "string",
                        "example": "true"
                    },
                    "count": {
                        "type": "number",
                        "example": 1
                    },
                    "percentage": {
                        "type": "number",
                        "example": 1
                    }
                }
            },
            "SuccessMessage": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    }
                }
            }
        }
    }
}